refactor: creates module date and transfers functions from module common
