+++ src/utils/common.js
@@ -1,7 +1,3 @@
-import dayjs from 'dayjs';
-
-const MIN_COEFFICIENT = 0;
-const MAX_COEFFICIENT = 30000;
 const MAX_VALUE = 10;
 
 export const getRandomIntegerNumber = (min, max) => {
@@ -14,19 +10,6 @@ export const getRandomArrayItem = (array) => {
   return array[randomItem];
 };
 
-export const getRandomDate = () => {
-  const targetDate = new Date();
-  const diffValue = getRandomIntegerNumber(MIN_COEFFICIENT, MAX_COEFFICIENT);
-
-  targetDate.setDate(targetDate.getDate() - diffValue);
-
-  return targetDate;
-};
-
 export const setValueFormat = (value) => {
   return value < MAX_VALUE ? `0${value}` : String(value);
-};
-
-export const getFullYear = (date) => {
-  return dayjs(date).format(`YYYY`);
 };
